refactor(crouton-builder-demo): extract spike-app logic into composables (#1240)
👤 For humans
spike-app.vue was the repo's #1 churn hotspot (83 commits on a 2100-line
file). Its script is now ~150 lines delegating to focused composables and
pure utils; the template and styles are byte-identical, so the page behaves
exactly the same.

🤖 For agents
- 8 composables (undo, snap-drag, focus-edit, pinch, node-ops, board-add,
  proposals, page-nav) — each called once from the page so singleton
  reactivity + provide() context are unchanged; shared seq counter stays in
  the page, passed as a nextSeq closure
- 2 utils: spike-board.ts (types + snap geometry; boardBounds/clearSpot now
  take nodes as a param) and spike-demo-data.ts (drawer/PAGES tables)
- Template/style blocks diff-verified byte-identical vs HEAD
- fallow hotspot score 57.1 → 30.6 (density 0.28 → 0.15); repo churn
  hotspots 1 → 0
